Feat refactor scroll utils
