+++ src/utils/nav.ts
@@ -1,0 +1,63 @@
+interface NavParams {
+    currentPath: string;
+    targetPath: string;
+    targetHash: string | undefined;
+    currentHash: string | undefined;
+    updateRoute: (route: string) => void;
+    handleScrollClick?: (hash: string) => void;
+}
+
+export const handleNavClick = (
+    path: string,
+    currentRoute: string | undefined,
+    updateRoute: (route: string) => void,
+    handleScrollClick?: (hash: string) => void
+): void => {
+    if (!currentRoute) {
+        return;
+    }
+    const [currentPath, currentHash] = currentRoute.split("#");
+    const [targetPath, targetHash] = path.split("#");
+    ifNav({ currentPath, targetPath, targetHash, currentHash, updateRoute });
+    elseNav({
+        currentPath,
+        targetPath,
+        targetHash,
+        currentHash,
+        updateRoute,
+        handleScrollClick,
+    });
+};
+
+function ifNav({ currentPath, targetPath, targetHash, currentHash, updateRoute }: NavParams): void {
+    if (currentPath !== targetPath) {
+        updateRoute(targetPath);
+        if (targetHash === undefined) {
+            return;
+        }
+        if (targetHash !== currentHash) {
+            updateRoute(`${targetPath}#${targetHash}`);
+        }
+    }
+}
+
+function elseNav({
+    currentPath,
+    targetPath,
+    targetHash,
+    currentHash,
+    updateRoute,
+    handleScrollClick,
+}: NavParams): void {
+    if (currentPath === targetPath) {
+        updateRoute(targetPath);
+        if (targetHash === undefined) {
+            handleScrollClick?.(`scroll-start`);
+        } else if (targetHash !== currentHash) {
+            handleScrollClick?.(targetHash);
+            updateRoute(`${targetPath}#${targetHash}`);
+        } else if (targetHash === currentHash) {
+            updateRoute(`${targetPath}#${targetHash}`);
+        }
+    }
+}
