+++ src/workers/scrollSmoothWorker.js
